+++ common/src/main/java/mca/entity/ai/brain/tasks/chore/ChoppingTask.java
@@ -43,7 +43,7 @@ public class ChoppingTask extends AbstractChoreTask {
 
     @Override
     protected boolean shouldKeepRunning(ServerWorld world, VillagerEntityMCA villager, long time) {
-        return shouldRun(world, villager) && villager.getHealth() == villager.getMaxHealth();
+        return shouldRun(world, villager);
     }
 
     @Override