+++ common/src/main/java/mca/entity/ai/brain/tasks/chore/HarvestingTask.java
@@ -62,7 +62,7 @@ public class HarvestingTask extends AbstractChoreTask {
 
     @Override
     protected boolean shouldKeepRunning(ServerWorld world, VillagerEntityMCA villager, long time) {
-        return shouldRun(world, villager) && villager.getHealth() == villager.getMaxHealth();
+        return shouldRun(world, villager);
     }
 
     @Override