+++ common/src/main/java/mca/entity/ai/brain/tasks/chore/HuntingTask.java
@@ -34,7 +34,7 @@ public class HuntingTask extends AbstractChoreTask {
 
     @Override
     protected boolean shouldKeepRunning(ServerWorld world, VillagerEntityMCA villager, long time) {
-        return shouldRun(world, villager) && villager.getHealth() == villager.getMaxHealth();
+        return shouldRun(world, villager);
     }
 
     @Override